remove old preview part, use new one
